+++ src/controllers/web/Oauth2ServerController.php
@@ -5,6 +5,7 @@ namespace rhertogh\Yii2Oauth2Server\controllers\web;
 use rhertogh\Yii2Oauth2Server\controllers\web\base\Oauth2BaseApiController;
 use rhertogh\Yii2Oauth2Server\controllers\web\server\Oauth2AccessTokenAction;
 use rhertogh\Yii2Oauth2Server\controllers\web\server\Oauth2AuthorizeAction;
+use rhertogh\Yii2Oauth2Server\controllers\web\server\Oauth2RevokeAction;
 use rhertogh\Yii2Oauth2Server\interfaces\controllers\web\Oauth2ServerControllerInterface;
 use yii\filters\VerbFilter;
 use yii\helpers\ArrayHelper;
@@ -22,6 +23,7 @@ class Oauth2ServerController extends Oauth2BaseApiController implements Oauth2Se
                 'actions' => [
                     static::ACTION_NAME_ACCESS_TOKEN => ['POST'],
                     static::ACTION_NAME_AUTHORIZE => ['GET', 'POST'],
+                    static::ACTION_NAME_REVOKE => ['POST'],
                 ],
             ],
         ]);
@@ -35,6 +37,7 @@ class Oauth2ServerController extends Oauth2BaseApiController implements Oauth2Se
         return [
             static::ACTION_NAME_ACCESS_TOKEN => Oauth2AccessTokenAction::class,
             static::ACTION_NAME_AUTHORIZE => Oauth2AuthorizeAction::class,
+            static::ACTION_NAME_REVOKE => Oauth2RevokeAction::class,
         ];
     }
 }